Datasets/evalplus/Java/src/test/java/com/example/HumanEval_56Test.java
package com.example;
import static org.junit.Assert.*;
import org.junit.Test;

public class HumanEval_56Test {

    @Test
    public void test_0() {
        assertTrue(HumanEval_56.correct_bracketing("<>"));
    }

    @Test
    public void test_1() {
        assertTrue(HumanEval_56.correct_bracketing("<<><>>"));
    }

    @Test
    public void test_2() {
        assertTrue(HumanEval_56.correct_bracketing("<><><<><>><>"));
    }

    @Test
    public void test_3() {
        assertTrue(HumanEval_56.correct_bracketing("<><><<<><><>><>><<><><<>>>"));
    }

    @Test
    public void test_4() {
        assertFalse(HumanEval_56.correct_bracketing("<<<><>>>>"));
    }

    @Test
    public void test_5() {
        assertFalse(HumanEval_56.correct_bracketing("><<>"));
    }

    @Test
    public void test_6() {
        assertFalse(HumanEval_56.correct_bracketing("<"));
    }

    @Test
    public void test_7() {
        assertFalse(HumanEval_56.correct_bracketing("<<<<"));
    }

    @Test
    public void test_8() {
        assertFalse(HumanEval_56.correct_bracketing(">"));
    }

    @Test
    public void test_9() {
        assertFalse(HumanEval_56.correct_bracketing("<<>"));
    }

    @Test
    public void test_10() {
        assertFalse(HumanEval_56.correct_bracketing("<><><<><>><>><<>"));
    }

    @Test
    public void test_11() {
        assertFalse(HumanEval_56.correct_bracketing("<><><<><>><>>><>"));
    }

    @Test
    public void test_12() {
        assertTrue(HumanEval_56.correct_bracketing("<<<>>>"));
    }

    @Test
    public void test_13() {
        assertFalse(HumanEval_56.correct_bracketing("><"));
    }

    @Test
    public void test_14() {
        assertFalse(HumanEval_56.correct_bracketing("<<>>><><><><>>>><"));
    }

    @Test
    public void test_15() {
        assertFalse(HumanEval_56.correct_bracketing(">>>>"));
    }

    @Test
    public void test_16() {
        assertTrue(HumanEval_56.correct_bracketing("<<<<>><>>><>"));
    }

    @Test
    public void test_17() {
        assertFalse(HumanEval_56.correct_bracketing("<<<<<>><><><><><><>><><>>>>><>><<<<<><><><<<<<<<<<<>>>>>>>>><<>>><><><><><><>><><><<><><><>>"));
    }

    @Test
    public void test_18() {
        assertFalse(HumanEval_56.correct_bracketing("<"));
    }

    @Test
    public void test_19() {
        assertTrue(HumanEval_56.correct_bracketing(""));
    }

    @Test
    public void test_20() {
        assertFalse(HumanEval_56.correct_bracketing("<<<<<<<>>><><><>>>>><<"));
    }
}
// End of Test
